refactor(spx-gui): extract asset library from AssetLibraryModal
- Extracted asset library from AssetLibraryModal.vue to AssetLibrary.vue
  to made modal a form of container for asset library
